chatApp/requirements.txt
# LangChain 核心
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# RAG 組件
chromadb>=0.5.0

# 向量嵌入
sentence-transformers>=2.2.0

# HTTP 客戶端（用於 MCP 通信）
httpx>=0.27.0

# 聯網搜索
tavily-python>=0.3.0

# 數據處理
python-dotenv>=1.0.0
pydantic>=2.0.0

# CLI 界面
rich>=13.0.0
prompt-toolkit>=3.0.0

# 日誌
loguru>=0.7.0
